Update kevin_courey.html
SoaringBow4 fixed all three videos by adding width and height values, as well as the "controls" attribute, which instructs the browser to include basic video controls: pause, play and skip. The text "Video not supported", was added as well and will only be displayed if the browser is unable to load the video.