Update matthew_ullmann.html
I added a div tag for the rating out of 10 with a paragraph explaining my reasons. I also added another div tag with a video link for the trailer of my game (Uncharted 4). I also have another div tag with the title as review with a paragraph giving my review of the game. I also have another div tag with the title of History of the Uncharted series with a paragraph explaining the history of my game (Uncharted 4). I also have another div tag with a title of system requirements and an unordered list of what is required to run the game. I finally added at the end of my html code for links for other website that you can inform yourself more on the topic.